feat: implement AdDableWidget and AdPopinRecommend components for improved ad display
