+++ packages/mirror-tv-next/components/story/_styles/ad-popin-recommend.module.scss
@@ -1,4 +1,5 @@
 .dailySelection {
+  max-width: fit-content;
   border-image: linear-gradient(
       180deg,
       #153047 35%,
@@ -12,15 +13,7 @@
     color: #153047 !important;
   }
 }
+
 .popinRecommend {
   min-height: 344px;
-}
-
-.dableWidgetLast {
-  min-height: 1050px;
-
-  @media (max-width: 767px) {
-    aspect-ratio: 343 / 1800;
-    min-height: 1800px;
-  }
 }
